Create class Computer
This class is taing care for the functions that the computer needs.
It has method #generate_number() - generates valid number for the computer.
We user this method in the __init__ to set the result from this method
to the class variable number.
In #make_a_guess() - return number that the computer is giving as a guess.